loader.c - fix errors for -m32 build (not tested)

+++ packer/linux_x86_64-userspace/src/loader.c
@@ -8,7 +8,7 @@
 
 bool pt_mode = true;
 
-static inline uint64_t perform_hypercall(uint64_t rbx, uint64_t rcx){
+static inline uint64_t perform_hypercall(uintptr_t rbx, uintptr_t rcx){
   if(pt_mode){
     return KAFL_HYPERCALL_PT(rbx, rcx);
   }
@@ -57,7 +57,7 @@ bool download_file(const char* filename, const char* dst){
 
   do{
     strcpy(stream_data, filename);
-    bytes = perform_hypercall(HYPERCALL_KAFL_REQ_STREAM_DATA, (uint64_t)stream_data);
+    bytes = perform_hypercall(HYPERCALL_KAFL_REQ_STREAM_DATA, (uintptr_t)stream_data);
 
     if(bytes == 0xFFFFFFFFFFFFFFFFUL){
       printf("HYPERVISOR: ERROR\n");
@@ -122,20 +122,20 @@ int main(int argc, char** argv){
 
     if(pt_mode){
       if(!download_file("hget", "hget")){
-        perform_hypercall(HYPERCALL_KAFL_USER_ABORT, (uint64_t)"Error: Can't get file 'hget'\n");
+        perform_hypercall(HYPERCALL_KAFL_USER_ABORT, (uintptr_t)"Error: Can't get file 'hget'\n");
       }
 
       if(!download_file("fuzz.sh", "fuzz.sh")){
-        perform_hypercall(HYPERCALL_KAFL_USER_ABORT, (uint64_t)"Error: Can't get file 'fuzz.sh'\n");
+        perform_hypercall(HYPERCALL_KAFL_USER_ABORT, (uintptr_t)"Error: Can't get file 'fuzz.sh'\n");
       }
     }
     else{
       if(!download_file("hget_no_pt", "hget")){
-        perform_hypercall(HYPERCALL_KAFL_USER_ABORT, (uint64_t)"Error: Can't get file 'hget_no_pt'\n");
+        perform_hypercall(HYPERCALL_KAFL_USER_ABORT, (uintptr_t)"Error: Can't get file 'hget_no_pt'\n");
       }
 
       if(!download_file("fuzz_no_pt.sh", "fuzz.sh")){
-        perform_hypercall(HYPERCALL_KAFL_USER_ABORT, (uint64_t)"Error: Can't get file 'fuzz_no_pt.sh'\n");
+        perform_hypercall(HYPERCALL_KAFL_USER_ABORT, (uintptr_t)"Error: Can't get file 'fuzz_no_pt.sh'\n");
       }
     }
